Added unit test.  Fixed a bunch of hygiene issues.
The test is compiling, but not fully working.
NOTE: THIS IS UNIT TEST, NOT A TYPICAL SAMPLE OF HOW TO USE THE LIBRARY, PLEASE DON'T CUT AND PASTE.  If we have time we'll add a better sample code with a proper server and client.
Also renamed some "SteamDatagram" things (which is the name used to refer to the relay network, not included with the opensource code) to "SteamnetworkingSockets".
Remove include of platform.h in a public header.  That's not supposed to be visible or needed outside of compiling the library.
Added a stub for ISteamNetworkingUtils.  The only thing that's available in the opensource library is the timestamp function.  In the Steamworks version there are utilities for reasoning about the relay network and estimating ping between two locations.