Add AES-GCM to crypto layer using OpenSSL EVP.
Also add a EVP_CIPHER_CTX_safe structure to manage allocating and cleaning up
EVP_CIPHER_CTX struct.  IN older versions of OpenSSL you could just declare an
EVP_CIPHER_CTX struct on the stack.  Now you have to go through a generic
allocation mechanism!?  SO we're going to be doing a heap allocation for every
single packet.  That is....terrible.  We should try to find a way around this,
even if it's gross, because doing a heap allocation per packet is not acceptible.