Refactor storage of crypto key pairs.
The basic change is that the base class will no longer store the raw key bytes in a buffer.  Instead, we have virtual functions to get and set the raw bytes, with the storage being specific to the key type and crypto provider.

The main purpose of these changes is to make it easier to get 25519 working in OpenSSL and bcrypt.

NOTE: This broke libsodium support.  We could fix it.  But I'm not sure we want to support libsodium or not.  If we support it, we should support it for all crypto needs, including per-packet encryption.  And that means accepting the 12-byte IV and 16-byte tag.  If that's what TLS does, there is some potential advantage to doing exactly what TLS does.  At the same time, I don't add 12 bytes to every packet lightly.

Reorganize 25519 files, made a placeholder for OpenSSL with a bunch of FIXMEs.  (Working on issue #46.)

Also, deleted RSA key support, since we don't need it for thiis project.