Fix typo using same state name twice.
I don't think this actually caused any problems with the current possible code flows, but it *reads* as if there is a bug.

I'm actually surprised it didn't generate a compiler warning.  Checking the same condition twice (when the condition is known to produce the same result in both cases, with no side effects) seems pretty suspicious.