Plumbing changes to be compatible with Steamworks SDK.
In Steamworks, we want to keep the number of odinary DLL imports to an
absolute minimum, because those come from steam_api.dll, which we do
not update.  (The developer needs to take a new one to update their SDK.)
The functions in the interface come from steamclient.dll, which we do
update frequently.

This resulted in the following changes:
* Printing and parsing of SteamNetworkingIPAddr and SteamNetworkingIdentity
  were moved to ISteamNetworkingUtils functions.  Most C++ code will still
  access them through the inline class methods, so they won't really care
  where they are implemented.
* Deleted some hashing functions.  This is not hard to write, and the
  definition of operator= gives you an example of what to do.
* SteamNetworkingMessage::Release will happen through a pointer, so that it
  can be dynamiclaly linked.

Other changes:
* Renamed an old name "STEAMDATAGRAMLIB" to "STEAMNETWORKINGSOCKETS" in
  a few places.
* Use STEAMNETWORKINGSOCKETS_ENABLE_SDR in a few places, to remove
  SDR-specific functionality.  (We might opensource the SDR support code
  in the future.)