Track reason for sending stats.
"Stats" really means anything in the protobuf blob.  There are many
different reasons why we might need to send a message: urgent ping needed
to confirm connectivity when we appear to be timing out, keepalive for idle
connection, when we actually have some stats we want to send to the peer,
or acks.  Relayed connections have even more reasons.  We were doing a good
job of displaying *what* we were sending, but looking at traces it was not
always clear to me *why* we weer sending.  And you cannot infer the why from
the what, because once we decide that we are going to send something, we then
opportunistically fit in as much as we can.