Fix memory leak.
Two changes to cleanup.  First, to it earlier, as soon as the connection
switches into a state where we know we will not need to send or receive
messages anymore.  (Don't wait until object destruction.)  Second, we
were not freeing up any queued messages in the m_messagesQueued linked
list.

Also fix a monir bug in the linger functionality.  The connection is still
"connected" for all internal purposes, it just is not visible to the API
anymore.  So we need our stats object to stay alive, and we expect our peer
to be replying in a timely manner.

Add some asserts and paranoia.