Refactoring SteamNetworkingMessage_t.
Working towards being able to use this object in a new API to send messages
efficiently.  (Issue #51.)

- Message objects are actually ref counted now.
- m_sender renamed to m_identityPeer
- Added a flags field.  Now you can know if the received message was sent
  reliable or not.
- Added inline protected destructor to enforce proper usage.
- Tweaked tons of comments
- When receiving a fragmented unreliable message, I found a place where we
  were doing more memory allocations and memcpying than necessary and
  optimized that.

Delete a #ifdef __cplusplus block.  This whole file is using tons of C++
features.  It was weird to have that one block have that #ifdef

Also: In ConnectionState_ProblemDetectedLocally and ConnectionState_FinWait,
don't try to take any action immediately.  Just trigger an immediate wake up
call, and we will take that action back out in the main loop, where we know
what's higher up on the call stack and what is safe to do.