Use public message structure internally.
- Get rid of SNPSendMessage_t, and just use CSteamNetworkingMessage for the send
message queue.
- Combined SteamNetworkingMessageQueue and SSNPSendMessageList.
- I realized that reference counting would not work, so I deleted the reference
count.  We need to be able to be able to modify the message structure immediately,
and so the original use case I had for reference counting didn't actually work.
(I had imagined that you could send a message and keep a reference to the message,
and then you could examine it.  But I realized this get complicated, and so I
abandoned this approach, and instead will use a more explicit mechanism.)
- Renamed some stuff since the purpose of those variables had changed.