Only end-to-end link state tracker does keepalives.
LinkStatsTrackerBase::GetNextThinkTimeInternal should not check for ordinary
keepalives, since NeedToSendStats doesn't deal with it.  Promote that check
to the end-to-end state tracker only.

This mismatch is causing asserts because we will ask if we need to send stats
right now, and the answer is "no", and then we ask when the next time is that
we need to send stats and the answer is "yesterday".