Simplify IThinker system.
After reviewing all existing use cases, I realized that all of them were
tolerant of being woken up earlier instead of late, and that all of the
complexity I was doing with tracking earliest and latest allowed wake up
times was not really getting anything.  Furthermore, the way the code was
written, forcing everybody to accept a 1ms late slack had the effect of
almost *always* waking up 1ms late, even if we could have woken up at the
right time.

So now each thinker just has a target wake up time, and we will not call
them before that, but we will call them as soon as possible after that.

I suspect that I could make this a lot better by using waitable timers
on Windows and the timerfd functions on linux.  I'd need a solution for
other POSIX systems, though.