Make it safe(r) to do static shutdown at any time.
Really, you ought to explicitly shut down the library.  However, if we do perform
static destruction without properly cleaning up the library (perhaps the app has
some panic fatal error function), at least nuke the thread, because it is almost
certainly going to crash if it does work after objects, such as the thinker
priority queue, get destructed.

There is an argument to be made that crashing is good when it alerts you to a
bug, and it's better to crash and simply not allow bad app behaviour.  In this
case, I do have at least one customer who wants it this way.