Rename some functions to fix naming conflicts.
The standalone library was defining some of the same symbols as the flat
interface in steam_api.dll, and this was leading to complications.  Renamed
those functions, and they are not part of the "flat interface", they are just
global internal functions functions.

Added wrappers in the flat interface with the same name to pass thorugh.  So
if the library is compiled with the flat interface (which we do in the
opensource version), no functions are removed and the flat interface stays the
same.