P2P / ICE imprvoements
This change includes many changes to improve P2P connections.  I didn't
bother replaying them one at a time from perforce.

Changed the plumbing of how data packets are sent, using context structures
more. The main goal of these changes is to remove the assumption that data
packets are always sent/received on the currently selected transport.  Also,
I removed two specialized end-to-end messages for ICE transport and SDR P2P
transport. Now all end-to-end packets are ordinary data packets (just without
any data).  This is good because, for UDP, they are encrypted.  (For SDR
connections we don't encrytpt them, because we want the relays to be able to
observe them and perhaps mutate them.)  And is reduces the number of different
packet types and codepaths that need to be supported.  This does have one
unfortunate downside: the receiver does not know which transport is used to
send dropped packets.  So, any out-of-band checks on alternate transports,
which do not contain data, and are much more likely to drop (since the
transport is not selectged) will be observed as a gap in the end-to-end
sequence number.  Oddly, the sender actually will have the more complete
picture (albeit after some delay), since he knows the outcome of every
single packet, and he knows what transport was used to send it.

Added a flag that can be used to indicate to the peer, "I am not using this
transport as my primary transport."  My protocol does not have an explicit
nomination framework.  Instead, any data packet that doesn't have this flag
is considered a nomination.

SNP now needs to remember what transport was used for each outgoing packet.

Added some plumbing so that I can collect data about how often ICE is
successful, how the routes compare to SDR, and if it fails, what progress
it made.

The ICE session interface offers more control over what kind of candidtes to
gather, and returns more detailed info about the candidates and route.  The
goal here is to allow a semi-private mode.  If you don't want to share your
public address with random Internet peers (to avoid getting DDoSed), but you
do want to get a LAN connection to any peers in your office / dorm (where LAN
broadcasts may not discover that you are on the same LAN), you can choose to
share only local addresses.  This seems oddd to only share the "private"
address while keeping your "public" address a secret -- but in this case
the "private" address is actually more sensitive, and the public address is
the one that script kiddies would need to boot you.  We are dealing with a
differet kind of thread here.  The option to go totally private and disable
ICE completely, or only share relayed candidates, or only share public
addressses, are all still available, if the threat model is different.