Clamp CalcConservativeTimeout
This was cuasing us to hang on to expected acks too long, instead
of timing them out.  Really, I don't expect anybody to use this code
in use cases where a ping time of 500ms+ is tolerable.  If the ping
time is that high, any values tuned inappropriately in this library
are probably the least of the problems.

If there is anybody from NASA thinking about using this code to talk
to the Mars rover, or maybe somebody working on RFC1149, well by all
means let me know and we can figure out a better solution.  Right now
I'm focused on terrestial communications over copper, radio, and fiber.