Prune webrtc dependencies.
The webrtc code is actually really well organized, with a layered approach.
We are accessing the code at alow level, and we really don't need any of
the higher levels.  If function-level linking is used, this problem
would already be solved and we really wouldn't need to do any work.

Unfortunately, right now webrtc_sdp.cc is a file that straddles layers.
The *functions* themselves are actually layered just fine.  But they are
in the same source file, and so when file-level linking is used, you cannot
use only the low level ICE-related functions.  You end up bringing in the
full SDP parsing and printing functions, which means you have to know about
media types, and codecs, etc.  I think the best way to cut this Gordian
knot is to....um...cut it.  See the hacked file for more info.

Right now we are not using DTLS, and so we still have some extra stuff in
the lib that could be removed.  But I'm not inclined to make those cuts right
now, because in the future we may want to use their DTLS code.  In fact,
we may end up needing to put some of this stuff back, if we want to use
RTP data channels, perhaps for compatibility with the browser.