Fix a subtle bug if ping increases suddenly
If latency spikes very suddenly and dramatically, we won't ever update our
estimate.

The problem is that we use the RTT to decide when we can forget about old
packets that we have marked as NACK because they timed out.  But this expiry
is based on our RTT estimate, which in this case is too small, perhaps
significantly too small.  And then, when those packets do eventually get
acked, we cannot update our RTT estimate, because we have forgotten about the
packet being acked!  So we will be stuck with an RTT estimate that is
permanently inaccurate.

The clue that something was wrong was in the connection test, I should
have noticed it.  The ping time goes from 0 (it's loopback) to 200ms
(because we are faking it).  And the ping time estimates being displayed
were still ~1ms, even though we were faking 200ms.

This is probably extremely rare in practice, but still it is relatively
catastrophic when it happens, so it's good to fix it.

P4:6402130