Split out some functionality of the CSteamNetworkingMessages into CMessagesEndPoint base class.
This is in preperation for making a FakeUDP port, which will derive from
CMessagesEndPoint.  Also introduced CMessagesEndPointSession, which is now
the base class of SteamNetworkingMessagesSession

As part of this work:
- Reworked how we link CMessagesEndPointSession and connections.  A connection
now has an explicit pointer to the owning CMessagesEndPointSession, and we have
a bit more explicit special-case plumbing for connection state changes, instead
of using the generic callback mechanism.  We can now unlink a connection from a
session, which means that we can delete the session, and leave the connection in
the finwait state to more gracefully clean up.

- Added CSteamNetworkingSockets::m_mapMessagesEndpointByVirtualPort, and kill all
messages endpoints in CSteamNetworkingSockets::KillConnections.

P4:6753554