Add CheckWaitForInitialRoutingReady
This has some code that was waiting to send the connect *request* signal until
routing is ready.  Breaking it out into a separate function makes it possible
to delay sending the ConnectOK signal, for the same reasons.

Also improve the logic so that we only wait on ICE if there are candidate
types that are allowed, but not yet gathered.  And on SDR, we will now wait
to send the ConnectOK until we have established the first sessions with the
SDR relays.  (SDR not relevant in the opensource code, tho.)

P4:6753628