Fix locking bug
In CSteamNetworkConnectionBase::ProcessDeletionList, if we have an ad-hoc
"messages" session, we'd take locks in a way that our lock hygiene code
didn't like, triggering an assert.  The assert might have been spurrious,
but since it's easy to write the code in a way that obeys the strict rules
that protect us from writing deadlocks and other bugs, I didn't bother
investigating whether a dealock was actually possible.

Also delete an old, inaccurate FIXME comment

P4:6753634