Rewrote reliable segment bookkeeping to work with multiple lanes.
Previously we were keeping the reliable segments that had been sent in two
non-overlapping std::maps: in-flight, and ready-for-retry.

Also, each in-flight packet would remember the reliable segments it contained
as begin/end pairs.  When a packet was acked or nacked, we would need to lookup
the range in the maps.

The new method puts all segments that have been sent in a linked list, and we
track the status of each segment.  The in-flight packets hold a reference to
them, so there is no more need to lookup.  Likewise, the retry list is just a
list of referneces into that list.

This is *much* more memory efficient:
- The in-flight references are each one short, instead of a 64-bit begin/end
  pair.
- And we use CUtlLinkedList, which stores the data in a contiguous array,
  instead of std::map, which makes tons of tiny allocations.

I went in a cirlce on the design for the data structures for this at least 4
or 5 times, probably spending 10+ hours on it until I finally got something I
didn't totally hate.  There was always a bubble under the carpet that I could
move around but not elimiate.  In this design, the bubble is one place where
we do do some linear searching: When we need to retry a segment, we have to
find the proper place to insert it into the retry queue.  This is because
reliable serialization assumes that the segments will be presented in order
of increasing stream position for a given lane.  However, at least this list
should usually be short.

In the typical case of relatively little packet loss and need for reliable
retry, this code is much, MUCH more efficient than the previous code with
memory (and probably CPU -- although I haven't measured it!).  I believe
that to fix the remaining bubble under the carpet I need to change the
reliable retry list to be a map, ordered by some segment serial number.
But I don't want to do that with a std::map.  So I think I might bring
back Valve's ordered map CUtlMap, which uses integers (shorts if you want)
as stable handles to the elements, and uses contiguous memory allocation.
That would be good for this case, and also optimize a few of the other maps
as well.  But I think this code is good enough for now.

Also: added per-lane stats for number of bytes pending in various states.

P4:6834549