Tweak how we limit the number of packets sent in a single burst.
(In one "think" call.)  This was previously a constant, k_nMaxPacketsPerThink,
which was 16.  Now it's derived from the UDP socket buffer size (which is
renamed to g_cbUDPSocketBufferSize and made public within the library).  This
is really an arbitrary method, but at least it has some remote connection to
science and scales with the library in some way.

P4:6877968,6877972