Don't "linger" if it isn't necessary.
When the "linger" mode is requested when closing a connection, check if we
actually need to do so.  If not, then immediately enter the FinWait state.
Otherwise, schedule an immediate wakeup.

The context here is that there's kind of sharp edge with closing connections
and then immediately terminating the app.  This is not as TCP connection, so
we cannot hand the connection over to the OS to clean it up.  The next change
will expand on this.

P4:6986508