Don't use std_vector for static vectors.
Some implementations of std::vector do dumb stuff and the default constructor
actually allocates memory.  (Uggggggg.)  This causes problems if an app wants
to set a custom memory allocator, because by the time they do that these
vectors will have already allocated memory, and then if that memory needs to be
reallocated, by the custom allocator, we run into problems.

This fixes #209

P4:7032464