Fix bug in ScheduledMethodThinker::EnsureMinScheduleTime.
The call to Cancel() cleared the current scheduled time, which made it
impossible for the subsequent call to IThinker::EnsureMinScheduleTime
to properly check if the new schedule time was newer.

This code is not supposed to be threadsafe regarding different threads
doing schedule operations on the same thinker at the same time, and if
the callback is about to be executed at the time we are modifying it,
it doesn't really matter exactly what happens as long as no data
structures are corrupted.  It is only supposed to be threadsafe with
regard to other thinkers and the thinker table.

This was causing a bug in the P2P transport selection code and was
caught by an assert in that code.

P4:7056505