Firm up some stuff wrt locking and object lifetimes.
Added some comments that clarify the rules around which locks must be held
to access which objects.

Added assertions to verify the appropriate locks are held, and provide better
feedback if a particular step takes a long time.

Move the sending of the initial auth message into shared code.

Deleted some code that was checking for a NULL m_pICESession.  That shoud
never happen based on the lifetimes of these objects.  It definitely
indicates a bug and I think it's fine to crash if it happens.

P4:7194292