Replace LinkStatsTracker boolean "passive" flag with an ActivityLevel enum.
Our LinkStatsTracker does a lot more than just track "stats", it is also has
the state machine for tracking timeouts, keepalives, etc.  In SDR, we have
some links that we maintain for backup purposes.  For those "backup"
connections, we *do* expect the peer to reply to us, and we do want to send
keepalives, only much less often.

Also, importantly, when we switch between states, we were clearing some
important timeout info for no good reason, and that made it complicated to
give more subtle diagnoses of exactly what was going wrong on the relay.

P4:8088044