Tweak ping calculations
We want to limit fluctuations caused by random noise on a small time scale,
when we are reciving many samples in quick succession.  This is common because
our protocol acks every packet and each ack can be used to measure the ping,
so when two peers are sending packets back and forth they basically get a
constant stream of ping measurements.

We also use the minimum as the aggregation function when many quick samples
are combined into a single sample.  See the comments for more info on why.

P4:8183131